// src/types/Vehicle.ts
export type Vehicle = {
  id?: number;
  vehicle_code?: number;
  licence_number?: number;
  vehicle_type_en?: string | null;
  vehicle_type_ar?: string | null;
  vehicle_model_ar?: string | null;
  vehicle_state?: string | null;
  vehicle_dept?: string | null;
  vehicle_model_en?: string | null;
  notes?: string | null;
};